refactor(requests): Restructured repo structure for requests page
