+++ src/app/[locale]/requests/clearingRequest/detail/[id]/components/ClearingRequestInfo.tsx
@@ -12,7 +12,7 @@
 import Link from 'next/link'
 import { useTranslations } from 'next-intl'
 import { ClearingRequestDetails } from '@/object-types'
-import styles from './../../../moderationRequest/[id]/moderationRequestDetail.module.css'
+import styles from '@/app/[locale]/requests/requestDetail.module.css'
 import { signOut, useSession } from 'next-auth/react'
