+++ src/app/[locale]/requests/clearingRequest/detail/[id]/page.tsx
