+++ src/app/[locale]/requests/components/ClosedClearingRequest.tsx
@@ -63,7 +63,7 @@ function ClosedClearingRequest() {
             formatter: ({ requestId }: { requestId: string; }) =>
                 _(
                     <>
-                        <Link href={`/requests/clearingRequest/${requestId}`} className='text-link'>
+                        <Link href={`/requests/clearingRequest/detail/${requestId}`} className='text-link'>
                             {requestId}
                         </Link>
                     </>
