+++ src/app/[locale]/requests/moderationRequest/[id]/components/ModerationRequestDetail.tsx
@@ -16,7 +16,7 @@ import { notFound, useRouter } from 'next/navigation'
 import { useCallback, useEffect, useRef, useState } from 'react'
 import { useTranslations } from 'next-intl'
 import { signOut, useSession } from 'next-auth/react'
-import styles from '../moderationRequestDetail.module.css'
+import styles from '@/app/[locale]/requests/requestDetail.module.css'
 import { Button, Col, Row, Tab, Card, Collapse } from 'react-bootstrap'
 import { ModerationRequestDetails } from '@/object-types'
 import ModerationRequestInfo from './ModerationRequestInfo'
