+++ src/app/[locale]/requests/moderationRequest/[id]/components/ModerationRequestInfo.tsx
@@ -12,7 +12,7 @@
 import Link from 'next/link'
 import { useTranslations } from 'next-intl'
 import { ModerationRequestDetails } from '@/object-types'
-import styles from '../moderationRequestDetail.module.css'
+import styles from '@/app/[locale]/requests/requestDetail.module.css'
 
 
 export default function ModerationRequestInfo({ data }: { data: ModerationRequestDetails }) {
